+++ packages/shared/sdk-client/src/context/addAutoEnv.ts
@@ -12,7 +12,8 @@ import {
 } from '@launchdarkly/js-sdk-common';
 
 import Configuration from '../configuration';
-import { getOrGenerateKey } from './getOrGenerateKey';
+import { getOrGenerateKey } from '../storage/getOrGenerateKey';
+import { namespaceForGeneratedContextKey } from '../storage/namespaceUtils';
 
 const { isLegacyUser, isSingleKind, isMultiKind } = internal;
 const defaultAutoEnvSchemaVersion = '1.0';
@@ -94,7 +95,8 @@ export const addDeviceInfo = async (platform: Platform) => {
 
   // Check if device has any meaningful data before we return it.
   if (Object.keys(device).filter((k) => k !== 'key' && k !== 'envAttributesVersion').length) {
-    device.key = await getOrGenerateKey('context', 'ld_device', platform);
+    const ldDeviceNamespace = namespaceForGeneratedContextKey('ld_device');
+    device.key = await getOrGenerateKey(ldDeviceNamespace, platform);
     device.envAttributesVersion = device.envAttributesVersion || defaultAutoEnvSchemaVersion;
     return device;
   }