+++ packages/sdk/akamai-edgekv/CHANGELOG.md
@@ -31,6 +31,16 @@ All notable changes to the LaunchDarkly SDK for Akamai Workers will be documente
     * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.1.1 to ^1.1.2
     * @launchdarkly/js-server-sdk-common bumped from ^2.2.1 to ^2.2.2
 
+## [1.1.12](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.1.11...akamai-server-edgekv-sdk-v1.1.12) (2024-08-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.1.11 to ^1.1.12
+    * @launchdarkly/js-server-sdk-common bumped from ^2.4.4 to ^2.4.5
+
 ## [1.1.11](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.1.10...akamai-server-edgekv-sdk-v1.1.11) (2024-07-01)