+++ packages/telemetry/node-server-sdk-otel/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.0.10](https://github.com/launchdarkly/js-core/compare/node-server-sdk-otel-v1.0.9...node-server-sdk-otel-v1.0.10) (2024-08-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @launchdarkly/node-server-sdk bumped from 9.5.0 to 9.5.1
+  * peerDependencies
+    * @launchdarkly/node-server-sdk bumped from >=9.4.3 to >=9.5.1
+
 ## [1.0.9](https://github.com/launchdarkly/js-core/compare/node-server-sdk-otel-v1.0.8...node-server-sdk-otel-v1.0.9) (2024-07-31)